Update README, add power curve easing functions, and refactor file structure.
