+++ README.md
@@ -1,3 +1,3 @@
 # inter.js
 
-Inter.js is intended to ease the creation of interactive javascript projects. The code is written in modern JavaScript, however, some parts of the library are not concerned with being easily understood, so please do explore `/doc`.
+Inter.js is intended to ease the creation of interactive javascript projects. The code is written in modern JavaScript. However, some parts of the library are more concerned with efficiency than being easily understood from source, so please do explore `/doc`.
