+++ src/animation.js
@@ -1,0 +1,7 @@
+class Animation {
+}
+class Keyframe {
+}
+class Interpolation {
+}
+export {Animation, Keyframe, Interpolation};
