+++ src/fsm.js
@@ -1,0 +1,3 @@
+class FSM {
+}
+export default FSM;
